+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added URL selection mode with the ability to open new links in Reade Mode
 - arXiv links now open the full paper instead of the abstract page, with equations converted to Unicode math
 - Add support for displaying AVIF images
+- Fix Mermaid diagrams not rendering
 
 **Comment Section**